code/requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
tinycss2>=1.2.1
lxml>=5.0.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
requests>=2.31.0
typing_extensions>=4.0.0

# CLI
click>=8.1.0

# GUI
customtkinter>=5.2.0
tkinterdnd2-universal>=1.7.0
reportlab>=4.0.0
Pillow>=10.1.0
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"

# Visualization (GUI + ML)
matplotlib>=3.8.0
numpy>=1.26.0

# Machine Learning (optional)
scikit-learn>=1.0.0
joblib>=1.2.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
